[TIP] Reorganize table folder within indicators module
