+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/actions_row_cell.tsx
@@ -7,10 +7,10 @@
 
 import React, { useContext, VFC } from 'react';
 import { EuiFlexGroup } from '@elastic/eui';
-import { InvestigateInTimelineButtonIcon } from '../../../timeline/components/investigate_in_timeline';
-import { Indicator } from '../../../../../common/types/indicator';
-import { OpenIndicatorFlyoutButton } from '../open_indicator_flyout_button/open_indicator_flyout_button';
-import { IndicatorsTableContext } from './context';
+import { InvestigateInTimelineButtonIcon } from '../../../../timeline/components/investigate_in_timeline';
+import { Indicator } from '../../../../../../common/types/indicator';
+import { OpenIndicatorFlyoutButton } from './open_flyout_button';
+import { IndicatorsTableContext } from '../contexts';
 
 const INVESTIGATE_TEST_ID = 'tiIndicatorTableInvestigateInTimelineButtonIcon';
