+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/cell_actions.tsx
@@ -7,12 +7,12 @@
 
 import React, { VFC } from 'react';
 import { EuiDataGridColumnCellActionProps } from '@elastic/eui/src/components/datagrid/data_grid_types';
-import { Indicator } from '../../../../../common/types/indicator';
-import { AddToTimelineCellAction } from '../../../timeline/components/add_to_timeline';
-import { FilterInCellAction } from '../../../query_bar/components/filter_in';
-import { FilterOutCellAction } from '../../../query_bar/components/filter_out';
-import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../utils/field_value';
-import type { Pagination } from '../../services/fetch_indicators';
+import { Indicator } from '../../../../../../common/types/indicator';
+import { AddToTimelineCellAction } from '../../../../timeline/components/add_to_timeline';
+import { FilterInCellAction } from '../../../../query_bar/components/filter_in';
+import { FilterOutCellAction } from '../../../../query_bar/components/filter_out';
+import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../../utils/field_value';
+import type { Pagination } from '../../../services/fetch_indicators';
 
 export const CELL_TIMELINE_BUTTON_TEST_ID = 'tiIndicatorsTableCellTimelineButton';
 export const CELL_FILTER_IN_BUTTON_TEST_ID = 'tiIndicatorsTableCellFilterInButton';
