+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/field_browser/field_browser.tsx
@@ -7,7 +7,7 @@
 
 import { BrowserField } from '@kbn/rule-registry-plugin/common';
 import { VFC } from 'react';
-import { useKibana } from '../../../../hooks/use_kibana';
+import { useKibana } from '../../../../../../hooks/use_kibana';
 
 export interface IndicatorsFieldBrowserProps {
   browserFields: Readonly<Record<string, Partial<BrowserField>>>;
