+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/open_flyout_button/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export * from './indicators_table';
+export * from './open_flyout_button';
