+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/open_flyout_button/open_flyout_button.stories.tsx
@@ -8,9 +8,9 @@
 import React from 'react';
 import { ComponentStory } from '@storybook/react';
 import { createKibanaReactContext } from '@kbn/kibana-react-plugin/public';
-import { mockUiSettingsService } from '../../../../common/mocks/mock_kibana_ui_settings_service';
-import { generateMockIndicator, Indicator } from '../../../../../common/types/indicator';
-import { OpenIndicatorFlyoutButton } from './open_indicator_flyout_button';
+import { mockUiSettingsService } from '../../../../../../common/mocks/mock_kibana_ui_settings_service';
+import { generateMockIndicator, Indicator } from '../../../../../../../common/types/indicator';
+import { OpenIndicatorFlyoutButton } from '.';
 
 export default {
   component: OpenIndicatorFlyoutButton,
