+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/open_flyout_button/open_flyout_button.test.tsx
@@ -7,9 +7,9 @@
 
 import React from 'react';
 import { render } from '@testing-library/react';
-import { BUTTON_TEST_ID, OpenIndicatorFlyoutButton } from './open_indicator_flyout_button';
-import { generateMockIndicator } from '../../../../../common/types/indicator';
-import { TestProvidersComponent } from '../../../../common/mocks/test_providers';
+import { BUTTON_TEST_ID, OpenIndicatorFlyoutButton } from '.';
+import { generateMockIndicator } from '../../../../../../../common/types/indicator';
+import { TestProvidersComponent } from '../../../../../../common/mocks/test_providers';
 
 const mockIndicator = generateMockIndicator();
