+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/components/open_flyout_button/open_flyout_button.tsx
@@ -8,7 +8,7 @@
 import React, { VFC } from 'react';
 import { EuiButtonIcon, EuiToolTip } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { Indicator } from '../../../../../common/types/indicator';
+import { Indicator } from '../../../../../../../common/types/indicator';
 
 export const BUTTON_TEST_ID = 'tiToggleIndicatorFlyoutButton';
