+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/contexts/context.ts
@@ -6,7 +6,7 @@
  */
 
 import { createContext, Dispatch, SetStateAction } from 'react';
-import { Indicator } from '../../../../../common/types/indicator';
+import { Indicator } from '../../../../../../common/types/indicator';
 
 export interface IndicatorsTableContextValue {
   expanded: Indicator | undefined;
