+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/hooks/use_toolbar_options.test.tsx
@@ -7,7 +7,7 @@
 
 import { TestProvidersComponent } from '../../../../../common/mocks/test_providers';
 import { renderHook } from '@testing-library/react-hooks';
-import { useToolbarOptions } from './use_toolbar_options';
+import { useToolbarOptions } from '.';
 
 describe('useToolbarOptions()', () => {
   it('should return correct value for 0 indicators total', () => {
