+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/table.stories.tsx
@@ -10,9 +10,9 @@ import { DataView } from '@kbn/data-views-plugin/common';
 import { mockIndicatorsFiltersContext } from '../../../../common/mocks/mock_indicators_filters_context';
 import { StoryProvidersComponent } from '../../../../common/mocks/story_providers';
 import { generateMockIndicator, Indicator } from '../../../../../common/types/indicator';
-import { IndicatorsTable } from './indicators_table';
+import { IndicatorsTable } from '.';
 import { IndicatorsFiltersContext } from '../../containers/indicators_filters/context';
-import { DEFAULT_COLUMNS } from './hooks/use_column_settings';
+import { DEFAULT_COLUMNS } from './hooks';
 
 export default {
   component: IndicatorsTable,
