+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/table/table.test.tsx
@@ -7,14 +7,10 @@
 
 import { act, render, screen } from '@testing-library/react';
 import React from 'react';
-import {
-  IndicatorsTable,
-  IndicatorsTableProps,
-  TABLE_UPDATE_PROGRESS_TEST_ID,
-} from './indicators_table';
+import { IndicatorsTable, IndicatorsTableProps, TABLE_UPDATE_PROGRESS_TEST_ID } from '.';
 import { TestProvidersComponent } from '../../../../common/mocks/test_providers';
 import { generateMockIndicator, Indicator } from '../../../../../common/types/indicator';
-import { BUTTON_TEST_ID } from '../open_indicator_flyout_button';
+import { BUTTON_TEST_ID } from './components/open_flyout_button';
 import { TITLE_TEST_ID } from '../flyout';
 import { SecuritySolutionDataViewBase } from '../../../../types';
