+++ x-pack/plugins/threat_intelligence/public/modules/indicators/indicators_page.tsx
@@ -8,7 +8,7 @@
 import React, { FC, VFC } from 'react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { IndicatorsBarChartWrapper } from './components/barchart';
-import { IndicatorsTable } from './components/indicators_table/indicators_table';
+import { IndicatorsTable } from './components/table';
 import { useIndicators } from './hooks/use_indicators';
 import { DefaultPageLayout } from '../../components/layout';
 import { useFilters } from '../query_bar/hooks/use_filters';
@@ -16,7 +16,7 @@ import { FiltersGlobal } from '../../containers/filters_global';
 import { useSourcererDataView } from './hooks/use_sourcerer_data_view';
 import { FieldTypesProvider } from '../../containers/field_types_provider';
 import { InspectorProvider } from '../../containers/inspector';
-import { useColumnSettings } from './components/indicators_table/hooks/use_column_settings';
+import { useColumnSettings } from './components/table/hooks';
 import { useAggregatedIndicators } from './hooks/use_aggregated_indicators';
 import { IndicatorsFilters } from './containers/indicators_filters';
 import { useSecurityContext } from '../../hooks/use_security_context';
